Prepare for deletion of newrelic.ts
