+++ .idea/codestream.iml
@@ -5,6 +5,7 @@
     <content url="file://$MODULE_DIR$">
       <excludeFolder url="file://$MODULE_DIR$/shared/agent/dist" />
       <excludeFolder url="file://$MODULE_DIR$/vscode/dist" />
+      <excludeFolder url="file://$MODULE_DIR$/dist" />
     </content>
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
